View does not define JUPYTER_DATA_DIR anymore, but the kernels and terminal require it.
The view does not define JUPYTER_DATA_DIR anymore because setting it to
$LCG_VIEW/share/jupyter, where the notebook extensions are, prevents the jupyter
server to start from an lxplus node configured with the view, since the directory
is not writable.

To solve this, we make this variable only available to the kernels and terminal,
so the user can enable correctly the nb extensions, otherwise they are not found
when running "jupyter nbextension enable ...".