Make _xsrf cookie a session cookie (#88)
The _xsrf cookie currently expires in 30 days, potentially in the middle
of a user session. This causes notebooks to fail to save until the page is
refreshed, which means the user loses unsaved work.

Since _xsrf cookie is set by the tornado server in jupyter-notebook. This commit
specifies two tornado settings to disable the expiration date being set in the cookie