Fix refreshing of EOS kerberos ticket in k8s
This commit is one of two parts of a fix for refreshing the EOS
kerberos ticket that is used by notebooks and terminal processes
in SWAN kubernetes, the other part being some changes in the
swan-cern chart. The changes of this commit rely on the
KRB5CCNAME_NB_TERM variable to be set in k8s to point to the
path where the notebook and terminal ticket is stored.