Install swandask extension v0.0.3
This extension is a wrapper to the dask-labextension extension.
It allows running the backend as a separate process, served via
jupyter-server-proxy, instead of a server extension.

We use it to configure the Dask environment, which we want to be
different than the process running Jupyter Lab.

The extension will automatically load the environment from the
default Python 3 kernel.

This extension needs to be installed after its dependency
dask-labextension in order to disable the server extension
automatically.