Generate self-signed certificate for Dask processes
This is done per user session, dynamically when the session starts.

The server certificate, private key and CA certificate can be used
to implement TLS in Dask clusters (e.g. via SwanHTCondorCluster).

Still missing: only generate the certificate if the user selects
"HTCondor cluster" from the form, which is not implemented yet.
In other words, the $CERN_HTCONDOR variable needs to be set
dynamically depending on the user form, instead of always like it
is done now.