Enable the use of a default Security object
This goes together with some changes in swan-daskcluster to be
able to automatically generate a Security object for Dask clients
that points to the certificates and keys we generate.

swan-daskcluster needs to be in the PYTHONPATH of notebooks and
terminals because it implements the loader who will create the
default Security object, and we point to that loader with a Dask
variable. Therefore, swan-daskcluster is installed in its own
directory so that it is added individually to the PYTHONPATH.