Align with other packages' line format for pattern matching in CI/CD
This will make it possible for the update-package job in the
github CI/CD of systemuser-image to match the swandaskcluster line
as it does with the SWAN extensions. Consequently, a new PR will
be automatically created in systemuser-image to update the
swandaskcluster version when there is a new swandaskcluster release.